refactor : decrease the function degree of freedom
